+++ pkg/api/api.go
@@ -258,8 +258,8 @@ func (a *API) RegisterQuerier(queryable storage.Queryable, engine *promql.Engine
 		querier.DummyTargetRetriever{},
 		querier.DummyAlertmanagerRetriever{},
 		func() config.Config { return config.Config{} },
-		map[string]string{},   // TODO: include configuration flags
-		v1.GlobalURLOptions{}, // TODO: what should this be?
+		map[string]string{}, // TODO: include configuration flags
+		v1.GlobalURLOptions{},
 		func(f http.HandlerFunc) http.HandlerFunc { return f },
 		func() v1.TSDBAdmin { return nil }, // Only needed for admin APIs.
 		false,                              // Disable admin APIs.