+++ pkg/querier/blocks_store_balanced_set.go
@@ -51,7 +51,8 @@ func (s *blocksStoreBalancedSet) starting(ctx context.Context) error {
 }
 
 func (s *blocksStoreBalancedSet) resolve(ctx context.Context) error {
-	return s.dnsProvider.Resolve(ctx, s.serviceAddresses)
+	_ = s.dnsProvider.Resolve(ctx, s.serviceAddresses)
+	return nil
 }
 
 func (s *blocksStoreBalancedSet) GetClientsFor(_ []*metadata.Meta) ([]storegatewaypb.StoreGatewayClient, error) {