+++ pkg/storegateway/bucket_stores_test.go
@@ -17,7 +17,6 @@ import (
 	"github.com/prometheus/prometheus/tsdb"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"github.com/thanos-io/thanos/pkg/block"
 	"github.com/thanos-io/thanos/pkg/store"
 	"github.com/thanos-io/thanos/pkg/store/storepb"
 	"github.com/weaveworks/common/logging"
@@ -27,8 +26,6 @@ import (
 	cortex_tsdb "github.com/cortexproject/cortex/pkg/storage/tsdb"
 	"github.com/cortexproject/cortex/pkg/util/flagext"
 )
-
-var _ block.MetadataModifier = &ReplicaLabelRemover{}
 
 func TestBucketStores_InitialSync(t *testing.T) {
 	userToMetric := map[string]string{